+++ src/api/viewsets/image/service.py
@@ -1,5 +1,6 @@
 from typing import Dict
 
+from api.viewsets import Image
 from api.viewsets.image.serializer import ImageSerializer
 
 
@@ -9,4 +10,4 @@ class ImageService:
     def create(image: Dict):
         image_serializer = ImageSerializer(data=image)
         image_serializer.is_valid(raise_exception=True)
-        image_serializer.save()
+        return Image.objects.create(**image_serializer.validated_data)